requirements.txt
django==5.1.0
django-rest-framework==0.1.0
django-storages==1.14.4
djangorestframework==3.15.2
celery==5.4.0
django-celery-beat==2.7.0
channels==3.0.0
channels-redis==4.2.0